requirements.txt
streamlit
python-docx
sentence-transformers
faiss-cpu
torch
numpy
google-generativeai>=0.7.2